Added booleOuts and "bits" value
Both the profile pages and account page now display booleOuts and a bits
value.  The booleOuts are currently the same ones that are displayed on
the bitstream, this will be changed, once the backend is ready.  The bit
value is currently 0, once again, this will change once the backend is
ready.